Nav Bar Setup
Got a navbar from bootstrap and edited to remove unnecessary components search as the search and dropdown for now. Edited the code so that it contains the needed components for my webpage eg the title and links of the page to match the pages I will have in my webpage.