Obtained the data for the products.
Obtained the data for the products sold at sunnybrae superette however it was in a xml format so I created a Script to conver the xml into a sql format. After transforming I have realised that lots of the products sold do not contain any prices, so i will have to cleanse the dataset prior to working on integrating the data onto the webpage.