Changes to the xml_to_sql
Made some changes to the xml_to_sql to ensure that there is a primary key involved and account for any issues with duplicates of primary keys.

Also ensured to delete all records where the price is 0 as the original data didnt provide the price for many of the products so I figured that i should remove them for now.