Product page layout fixed
Fixed the page so that the products are now appearing in proper boxes as opposed to earlier when it was appearing all below each other. Turns out i just had a closing div tag outside the for loop as opposed to inside it.

Also made changes so that the for loop is using the products from the page rather than all the products.