Added a load more button
Added a load more button, so that users can load the content as they want to. It will help to reduce load times as well.

Only issue is that the page number bar is still there and still a very long bar. So still need to fix it so that the bar isnt taking up the whole page.