Cart updates
Updated the following:

Models.py - updated to return price, quantity and totals for orders in cart.

Views.py - updated so that the order can be rendered in the cart for authenticated users so far.

cart.html - edited so that instead of displaying the dummy variables i had earlier it will now be responsive to the dummy user and the order that was created.