Load More button working.
Decided to add the Javascript function by function as opposed to everything straight away. So for more ensuring that the Load-More btn is working.

There was an issue with it and it wasnt working, i thought maybe it was due to the Dom not loading properly however it was because i didnt close the <script> link in main.html