Rest of the jsMoved into main.js
Moved the rest of the js code into main.js.

Add Button is working as it is outputting the product and action
Same with the arrows to change quantity on the cart page

however even though the buttons are working the quantity is not being updated on the cart page. So updateUserOrder most likely will need to be checked and updated.