Register and Login
Register and Login function are both working, user can be registered and they can log in, once logged in the logout button will appear as well. Only issue is that it doesnt create a customer instance for the user when they log in so they cannot interact with the cart page.