Created a Search button
The search button has been created and it works, although the search isn't as good as i want it to be so will need to touch up on it so that its more stricter.

Issue = There is one issue in that when I click the load more button it is loading items when havent been searched for, it just loads the default items on the page. So will need to find a fix for that as well.