Fixed issue of Pagination not working with the category's
When a category is selected from the drop down the pagination doesn't work and resets so fixed it via using the previous fix for the search as base. Was about to re code then realised that query_dict can just be used based on if the user is searching or using the category drop down.