adjusted the empty cart issue
When i was checking the order, even if the cart had items the message for no items in cart was popping up so adjusted for the time being to check if customer exists and if not then it will know the cart is empty. As by that point if there is no customer then there is no order.